Refactor files for correct folders
