+++ Python/Staircase.py
