+++ Python/defineaword.py
